requirements.txt
Django
djangorestframework
overrides
django-extensions